emulator-core: move the shared triangle setup into gpu/raster.rs
Pure code motion. raster.rs already held the hardware extent rule and
its header pointed at the rasterizer 'now living in gpu.rs'; the
setup (tri_raster_setup, the plane/span helpers, for_each_tri_pixel)
now lives beside it, and gpu.rs re-exports the public pieces so
psx-gpu-render's paths are unchanged. gpu.rs drops ~226 lines of
free functions that sat above the Gpu struct.

Verified: 433 + 126 tests green, canonical alttp 500M-step hashes
bit-identical.
